update deployment scripts
